feat: support Sections layout (#12)
* Fix indentation and trainling whitespace

* Add error on missing/wrong entities

* Cleanup css, remove unnecessary settings

* Add support for sizing in "Sections" layout

The "Sections" layout does not support autosizing of cards, so we have
to fix the size of the card to the configured size or there will be
overlaps or holes in the layout.

This does not remove the ability to resize the card in the "Masonry"
layout.

* Use default font size

* Update margins to work with different card sizes

With this margins the card work with odd heights without cutting of a
departure.

* Minimize bottom margin for Masonry layout